modularized card, deck, pile classes
